fixing alembic revisions. ETL runs successfully.
