--- alembic/versions/eff64db6b54f_mv_fermentation_update.py
@@ -1,149 +1,0 @@
-"""mv_fermentation_update
-
-Revision ID: eff64db6b54f
-Revises: d2b6b2a7c9d1
-Create Date: 2026-04-24 14:57:13.151606
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'eff64db6b54f'
-down_revision: Union[str, Sequence[str], None] = 'd2b6b2a7c9d1'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    op.add_column('method', sa.Column('duration', sa.Float(), nullable=True))
-
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_resource_strain")
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_product_name")
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_strain_name")
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_county")
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_geoid")
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_resource_id")
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_id")
-
-    op.execute("DROP MATERIALIZED VIEW IF EXISTS data_portal.mv_biomass_fermentation CASCADE")
-    op.execute("""
-        CREATE MATERIALIZED VIEW data_portal.mv_biomass_fermentation AS
-        SELECT row_number() OVER (ORDER BY fermentation_record.resource_id, location_address.geography_id, strain.name, pm.name, em.name, bm.name, parameter.name, unit.name) AS id,
-               fermentation_record.resource_id,
-               resource.name AS resource_name,
-               location_address.geography_id AS geoid,
-               place.county_name AS county,
-               strain.name AS strain_name,
-               pm.name AS pretreatment_method,
-               em.name AS enzyme_name,
-               bm.name AS bioconversion_method,
-               coalesce(pm.duration, em.duration, bm.duration) AS elapsed_time,
-               parameter.name AS product_name,
-               avg(observation.value) AS avg_value,
-               min(observation.value) AS min_value,
-               max(observation.value) AS max_value,
-               stddev(observation.value) AS std_dev,
-               count(*) AS observation_count,
-               unit.name AS unit
-        FROM fermentation_record
-        JOIN resource ON fermentation_record.resource_id = resource.id
-        LEFT OUTER JOIN prepared_sample ON fermentation_record.prepared_sample_id = prepared_sample.id
-        LEFT OUTER JOIN field_sample ON prepared_sample.field_sample_id = field_sample.id
-        LEFT OUTER JOIN location_address ON field_sample.sampling_location_id = location_address.id
-        LEFT OUTER JOIN place ON location_address.geography_id = place.geoid
-        LEFT OUTER JOIN strain ON fermentation_record.strain_id = strain.id
-        LEFT OUTER JOIN method AS pm ON fermentation_record.pretreatment_method_id = pm.id
-        LEFT OUTER JOIN method AS em ON fermentation_record.eh_method_id = em.id
-        LEFT OUTER JOIN method AS bm ON fermentation_record.method_id = bm.id
-        JOIN observation ON lower(observation.record_id) = lower(fermentation_record.record_id)
-        JOIN parameter ON observation.parameter_id = parameter.id
-        LEFT OUTER JOIN unit ON observation.unit_id = unit.id
-        WHERE fermentation_record.qc_pass != 'fail'
-        GROUP BY fermentation_record.resource_id,
-                 resource.name,
-                 location_address.geography_id,
-                 place.county_name,
-                 strain.name,
-                 pm.name,
-                 em.name,
-                 bm.name,
-                 coalesce(pm.duration, em.duration, bm.duration),
-                 parameter.name,
-                 unit.name
-    """)
-
-    op.execute("""CREATE UNIQUE INDEX idx_mv_biomass_fermentation_id ON data_portal.mv_biomass_fermentation (id)""")
-    op.execute("""CREATE INDEX idx_mv_biomass_fermentation_resource_id ON data_portal.mv_biomass_fermentation (resource_id)""")
-    op.execute("""CREATE INDEX idx_mv_biomass_fermentation_geoid ON data_portal.mv_biomass_fermentation (geoid)""")
-    op.execute("""CREATE INDEX idx_mv_biomass_fermentation_county ON data_portal.mv_biomass_fermentation (county)""")
-    op.execute("""CREATE INDEX idx_mv_biomass_fermentation_strain_name ON data_portal.mv_biomass_fermentation (strain_name)""")
-    op.execute("""CREATE INDEX idx_mv_biomass_fermentation_product_name ON data_portal.mv_biomass_fermentation (product_name)""")
-    op.execute("""CREATE INDEX idx_mv_biomass_fermentation_resource_strain ON data_portal.mv_biomass_fermentation (resource_id, strain_name)""")
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_resource_strain")
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_product_name")
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_strain_name")
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_county")
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_geoid")
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_resource_id")
-    op.execute("DROP INDEX IF EXISTS data_portal.idx_mv_biomass_fermentation_id")
-
-    op.execute("DROP MATERIALIZED VIEW IF EXISTS data_portal.mv_biomass_fermentation CASCADE")
-    op.execute("""
-        CREATE MATERIALIZED VIEW data_portal.mv_biomass_fermentation AS
-        SELECT row_number() OVER (ORDER BY fermentation_record.resource_id, location_address.geography_id, strain.name, pm.name, em.name, parameter.name, unit.name) AS id,
-               fermentation_record.resource_id,
-               resource.name AS resource_name,
-               location_address.geography_id AS geoid,
-               place.county_name AS county,
-               strain.name AS strain_name,
-               pm.name AS pretreatment_method,
-               em.name AS enzyme_name,
-               parameter.name AS product_name,
-               avg(observation.value) AS avg_value,
-               min(observation.value) AS min_value,
-               max(observation.value) AS max_value,
-               stddev(observation.value) AS std_dev,
-               count(*) AS observation_count,
-               unit.name AS unit
-        FROM fermentation_record
-        JOIN resource ON fermentation_record.resource_id = resource.id
-        LEFT OUTER JOIN prepared_sample ON fermentation_record.prepared_sample_id = prepared_sample.id
-        LEFT OUTER JOIN field_sample ON prepared_sample.field_sample_id = field_sample.id
-        LEFT OUTER JOIN location_address ON field_sample.sampling_location_id = location_address.id
-        LEFT OUTER JOIN place ON location_address.geography_id = place.geoid
-        LEFT OUTER JOIN strain ON fermentation_record.strain_id = strain.id
-        LEFT OUTER JOIN method AS pm ON fermentation_record.pretreatment_method_id = pm.id
-        LEFT OUTER JOIN method AS em ON fermentation_record.eh_method_id = em.id
-        JOIN observation ON lower(observation.record_id) = lower(fermentation_record.record_id)
-        JOIN parameter ON observation.parameter_id = parameter.id
-        LEFT OUTER JOIN unit ON observation.unit_id = unit.id
-        WHERE fermentation_record.qc_pass != 'fail'
-        GROUP BY fermentation_record.resource_id,
-                 resource.name,
-                 location_address.geography_id,
-                 place.county_name,
-                 strain.name,
-                 pm.name,
-                 em.name,
-                 parameter.name,
-                 unit.name
-    """)
-
-    op.execute("""CREATE UNIQUE INDEX idx_mv_biomass_fermentation_id ON data_portal.mv_biomass_fermentation (id)""")
-    op.execute("""CREATE INDEX idx_mv_biomass_fermentation_resource_id ON data_portal.mv_biomass_fermentation (resource_id)""")
-    op.execute("""CREATE INDEX idx_mv_biomass_fermentation_geoid ON data_portal.mv_biomass_fermentation (geoid)""")
-    op.execute("""CREATE INDEX idx_mv_biomass_fermentation_county ON data_portal.mv_biomass_fermentation (county)""")
-    op.execute("""CREATE INDEX idx_mv_biomass_fermentation_strain_name ON data_portal.mv_biomass_fermentation (strain_name)""")
-    op.execute("""CREATE INDEX idx_mv_biomass_fermentation_product_name ON data_portal.mv_biomass_fermentation (product_name)""")
-    op.execute("""CREATE INDEX idx_mv_biomass_fermentation_resource_strain ON data_portal.mv_biomass_fermentation (resource_id, strain_name)""")
-
-    op.drop_column('method', 'duration')
